+++ MonkeyLoader.Resonite.Data/MonkeyLoader.Resonite.Data.csproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="MonkeyLoader" Version="0.22.6-beta" />
+    <PackageReference Include="MonkeyLoader" Version="0.22.7-beta" />
     <PackageReference Include="PolySharp" Version="1.14.1">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
