+++ MonkeyLoader.Resonite.Integration/DataFeeds/Settings/SettingsHelpers.cs
@@ -1,5 +1,7 @@
-﻿using FrooxEngine;
+﻿using EnumerableToolkit;
+using FrooxEngine;
 using MonkeyLoader.Configuration;
+using MonkeyLoader.Logging;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -9,10 +11,23 @@ namespace MonkeyLoader.Resonite.DataFeeds.Settings
     public static class SettingsHelpers
     {
         public const string ConfigKeyChangeLabel = "Settings";
+        public const string ConfigSections = "ConfigSections";
         public const string EarlyMonkeys = "EarlyMonkeys";
+        public const string MetaData = "MetaData";
         public const string MonkeyLoader = "MonkeyLoader";
         public const string Monkeys = "Monkeys";
         public const string MonkeyToggles = "MonkeyToggles";
+        public const string ResetConfig = "ResetConfig";
+        public const string SaveConfig = "SaveConfig";
+        private static readonly Dictionary<SettingsDataFeed, SettingsViewData> _settingsViewsByFeed = [];
+
+        private static Logger Logger => MonkeyLoaderRootCategorySettingsItems.Logger;
+
+        public static SettingsViewData GetViewData(this SettingsDataFeed dataFeed)
+            => _settingsViewsByFeed.GetOrCreateValue(dataFeed, () => CreateViewData(dataFeed));
+
+        public static void InitBase(this DataFeedItem item, IReadOnlyList<string> path, IReadOnlyList<string> groupKeys, IDefiningConfigKey configKey)
+            => item.InitBase(configKey.FullId, path, groupKeys, configKey.GetLocaleString("Name"), configKey.GetLocaleString("Description"));
 
         public static void SetupConfigKeyField<T>(this IField<T> field, IDefiningConfigKey<T> configKey)
         {
@@ -25,6 +40,22 @@ namespace MonkeyLoader.Resonite.DataFeeds.Settings
             }
 
             field.SyncWithConfigKey(configKey, ConfigKeyChangeLabel);
+        }
+
+        private static SettingsViewData CreateViewData(SettingsDataFeed dataFeed)
+        {
+            static void OnDestroyed(IDestroyable destroyable)
+            {
+                destroyable.Destroyed -= OnDestroyed;
+                _settingsViewsByFeed.Remove((SettingsDataFeed)destroyable);
+
+                Logger.Debug(() => $"Removed ViewData for SettingsDataFeed ({destroyable.ReferenceID})");
+            }
+
+            var viewData = new SettingsViewData(dataFeed);
+            dataFeed.Destroyed += OnDestroyed;
+
+            return viewData;
         }
     }
 }
