+++ MonkeyLoader.Resonite.Integration/MonkeyLoader.Resonite.Integration.csproj
@@ -54,7 +54,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="MonkeyLoader" Version="0.22.6-beta" />
+    <PackageReference Include="MonkeyLoader" Version="0.22.7-beta" />
     <PackageReference Include="PolySharp" Version="1.14.1">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
